requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.0
certifi==2023.11.17
distro==1.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
litellm==1.17.5
openai==1.7.2
pyyaml==6.0.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
tqdm==4.66.1
typing_extensions==4.9.0
streamlit==1.32.0
pandas==2.1.4
matplotlib==3.8.3
plotly==5.19.0
